581 move is_smooth from number.py
